requirements.txt
aiogram==3.13.1
pynput==1.7.7
python-dotenv==1.0.1
